# Break out inner loop of optimize_layout function