...so numba can parallelize it effectively across all cores of a machine.

During benchmarking, I found that numba was not parallelizing `optimize_layout`, which is the most compute intensive part of UMAP. By extracting the code that runs for each epoch into its own function and decorating with numba's `parallel=True` I found that the computation can take advantage of all cores in a machine.

For example, running locally on a 4-core Mac I found one benchmark (of `optimize_layout`) went from ~110s to ~45s. And on a 16-core Linux box, the UMAP step of Scanpy went from ~206s to ~73s (for 130K rows).