@pep8speaks 2019-05-03
Hello @tomwhite! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`umap/umap_.py`](https://github.com/lmcinnes/umap/blob/8314d5d192bd3a7a21b1037a844c96633f55ff7c/umap/umap_.py):

> [Line 857:80](https://github.com/lmcinnes/umap/blob/8314d5d192bd3a7a21b1037a844c96633f55ff7c/umap/umap_.py#L857): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (86 > 79 characters)