@jlmelville 2019-05-08
@tomwhite, do you get the same coordinates values on multiple runs with the same random number seed?