@tomwhite 2019-05-08
@jlmelville, no, the coordinates change even with the same seed. In `_optimize_layout_single_epoch` there is a shared `rng_state` object and with multiple threads this state may not be updated determinstically. In `pynndescent` there is a `seed_per_row` option which computes the seed based on the row number, and uses a `rng_state` object per thread, but this only works because the threading there is done manually, not via numba's `parallel=True`. Unless there's a way of doing something similar with `parallel=True`, perhaps we should take the same approach here.