@lmcinnes 2019-05-08
Sounds good. I will merge this into the 0.4dev branch if possible so I can give it a little more testing before actually getting released into the wild.