@jlmelville 2019-05-09
My sense developing `uwot` was that a UMAP implementation that doesn't give reproducible results with a fixed seed would get a lot of push-back from scientific users (which I understand, being a lapsed scientist myself). Is there a way to this off if needed?

Personally, I only use UMAP for visualization, so I look forward to a faster optimization.