@tomwhite 2019-05-23
I had a bit more of a look at Numba's parallel implementation, and while it has [per-thread random state](https://github.com/numba/numba/commit/c7a49d148d9934fc0b781c704ee164d1e99c538d), does not seed these states, so it's not possible to make the calculation reproducible in this case.

So one way to fix this would be to use `parallel=True` only if UMAP's `random_state` is None, otherwise it would drop back to the sequential version for reproducibility, perhaps with a warning. I've prototyped this successfully.

Another related issue: it would be good to use NumPy's random library for random number generation, instead of the `tau_rand_int` function. This is because Numba creates per-thread state for random number generation, whereas `tau_rand_int` relies on a shared state and its unclear if that is safe for separate threads. Are there any objections to replacing `tau_rand_int` with `np.random`?