@tomwhite 2019-05-30
Some notes after further investigation:

1. UMAP's `tau_rand` random number generation doesn't play well with Numba's parallelization, since the latter doesn't provide a way to have per-thread state (there's an open issue though: https://github.com/numba/numba/issues/3656). Having a random number state shared between threads is potentially unsafe, and causes problems with reproducibility.
2. Conversely, Numba's `np.random` random number generation doesn't play well with `joblib`, since Numba doesn't expose the per-thread state (it's only available to its own parallelization) and [it doesn't support creating random state instances](https://numba.pydata.org/numba-doc/dev/reference/numpysupported.html#random). So, again, having a single global state shared between threads is potentially unsafe and precludes reproducibility.
3. Using Numba's `np.random` random number generation with Numba's parallelization would be safe (since internally it uses per-thread states), but these states cannot be seeded, which impacts reproducibility. Also, the threading support in pynndescent uses `joblib`, since it can't be expressed using Numba parallelization, and it probably makes sense to use the same random number generators in both projects (`tau_rand`).
4. UMAP's `tau_rand` random number generation _does_ work well with `joblib`, since the latter allows management of per-thread state, which means that the code is threadsafe and reproducible. (Reproducibility can be achieved by seeding each random state with a chunk index.) The downside is that it is a bit more work to add - it's not just a question of adding a Numba decorator.

I've implemented 4 in this branch: https://github.com/tomwhite/umap/tree/embedding_optimization_joblib. It achieves a comparable speedup to this PR, and the output of MNIST looks very similar.

Having got to this point, I'm not sure if this should be merged though! The reason is that the inner loop of `optimize_layout` does non-local updates of the embedding matrix (i.e. not just row `i`), so when multiple threads are running in parallel their updates could interfere, so that rows are not updated atomically. (Note that this observation is unrelated to points 1-4 above, and would be a problem under any of those implementations.) This may not be a problem in practice - and indeed this is reminiscent of [Hogwild!](https://papers.nips.cc/paper/4390-hogwild-a-lock-free-approach-to-parallelizing-stochastic-gradient-descent), although Hogwild! depends on atomic updates of single components of a weight vector, which is not what is happening here.

Perhaps some of the other approaches for optimizing SGD in #37 would be better.