@jlmelville 2019-05-30
The non-local update is how LargeVis works, so there is a precedent. I have implemented this scheme in uwot, and [visually](https://jlmelville.github.io/uwot/fast-sgd.html), it doesn't make much difference, so for visualizations it is a useful option to have. But it is not on by default in the name of preserving reproducibility.